+++ esti/hooks_test.go
@@ -13,10 +13,8 @@ import (
 	"text/template"
 	"time"
 
-	"github.com/cenkalti/backoff/v4"
 	"github.com/stretchr/testify/require"
 	"github.com/treeverse/lakefs/pkg/api/apigen"
-	"github.com/treeverse/lakefs/pkg/api/apiutil"
 )
 
 //go:embed action_files/*.yaml
@@ -74,28 +72,6 @@ func TestHooksSuccess(t *testing.T) {
 		valIdx := len(hooksTestData.data) - (i + 1)
 		require.Equal(t, hooksTestData.data[valIdx].EventType, run.EventType)
 	}
-}
-
-func WaitForListRepositoryRunsLen(ctx context.Context, t *testing.T, repo, ref string, l int) *apigen.ActionRunList {
-	var runs *apigen.ActionRunList
-	bo := backoff.NewExponentialBackOff()
-	bo.MaxInterval = 5 * time.Second
-	bo.MaxElapsedTime = 30 * time.Second
-	listFunc := func() error {
-		runsResp, err := client.ListRepositoryRunsWithResponse(ctx, repo, &apigen.ListRepositoryRunsParams{
-			Commit: apiutil.Ptr(ref),
-		})
-		require.NoError(t, err)
-		runs = runsResp.JSON200
-		require.NotNil(t, runs)
-		if len(runs.Results) == l {
-			return nil
-		}
-		return fmt.Errorf("run results size: %d", len(runs.Results))
-	}
-	err := backoff.Retry(listFunc, bo)
-	require.NoError(t, err)
-	return runs
 }
 
 func testCommitMerge(t *testing.T, ctx context.Context, repo string) {
